test: Compare integers by same signs
`BOOST_*` allows for comparison of integers with different signs. This
is problematic in the case of overflow and underflow, and should not be
valid for a unit test framework.

+++ src/test/skiplist_tests.cpp
@@ -137,7 +137,7 @@ BOOST_AUTO_TEST_CASE(findearliestatleast_test)
     for (unsigned int i=0; i<10000; ++i) {
         // Pick a random element in vBlocksMain.
         int r = m_rng.randrange(vBlocksMain.size());
-        int64_t test_time = vBlocksMain[r].nTime;
+        unsigned int test_time = vBlocksMain[r].nTime;
         CBlockIndex* ret = chain.FindEarliestAtLeast(test_time, 0);
         BOOST_CHECK(ret->nTimeMax >= test_time);
         BOOST_CHECK(((ret->pprev==nullptr) || ret->pprev->nTimeMax < test_time));
@@ -184,11 +184,11 @@ BOOST_AUTO_TEST_CASE(findearliestatleast_edge_test)
     BOOST_CHECK(!chain.FindEarliestAtLeast(0, 9));
 
     CBlockIndex* ret1 = chain.FindEarliestAtLeast(100, 2);
-    BOOST_CHECK(ret1->nTimeMax >= 100);
+    BOOST_CHECK(ret1->nTimeMax >= 100U);
     BOOST_CHECK(ret1->nHeight == 2);
     BOOST_CHECK(!chain.FindEarliestAtLeast(300, 9));
     CBlockIndex* ret2 = chain.FindEarliestAtLeast(200, 4);
-    BOOST_CHECK(ret2->nTimeMax >= 200);
+    BOOST_CHECK(ret2->nTimeMax >= 200U);
     BOOST_CHECK(ret2->nHeight == 4);
 }
